Test Suite: Change Package and Id to recommended in Constructor VI
This changes the Package and Id terminals of the Constructor VI from
required to recommended and initializes them with Package = "" and Id =
-1, respectively, by default.

Package and Id are only required when test suites are defined in an
aggregated file. For non-aggregated files, these attributes are not
used.

The 'package' attribute is omitted when Package is an empty string.
The 'id' attribute is omitted when Id is a negative value.