Add support for non-aggregated documents
This enables support for creating non-aggregated documents. In
non-aggregated documents, the `testsuites` element is omitted and only a
single `testsuite` is defined.

To make this possible, the caller can now connect either a document or
an element to the `To XML` VI, which then either adds the nodes to the
parent element or the document.

Note that the current implementation also allows any other entity
(Property, Test Case) to be serialized as root element. This, however,
is not allowed by the specification.